add accept and reject friends request functions

+++ controllers/friends_controller.js
@@ -5,15 +5,29 @@ const { validationResult } = require("express-validator");
 module.exports = {
     addFriend: async (req, res) => {
         try {
-            const to = req.body.to;
+            const to = req.body.frndUserName;
             const userId = req.userId;
             const errors = validationResult(req);
 
-            if (!errors.isEmpty())
+            if (!errors.isEmpty()) {
                 return res.status(400).json({
                     errors: errors.array(),
                     message: "Please enter valid user name",
                 });
+            }
+
+            const frnd = await USERS.findOne({ userName: to });
+            if (!frnd) {
+                return res.status(404).json({
+                    message: "No user found Please check the user name",
+                });
+            }
+
+            if (frnd._id.equals(userId)) {
+                return res.status(409).json({
+                    message: "You can't send frnd request to user self",
+                });
+            }
 
             await FRIEND_REQ.create({
                 from: userId,
@@ -67,18 +81,62 @@ module.exports = {
                 USERS.findByIdAndUpdate(friendId, {
                     $push: { friends: userId },
                 }),
-                FRIEND_REQ.findByIdAndDelete(requestId)
+                FRIEND_REQ.findByIdAndDelete(requestId),
             ];
 
-            await Promise.all(promises)
+            await Promise.all(promises);
             res.status(200).json({
                 message: "Friend request accepted successfully",
             });
-
         } catch (error) {
             res.status(500).json({
                 error: error.message,
                 message: "Internal server error",
+            });
+        }
+    },
+
+    rejectFriendRequest: async (req, res) => {
+        try {
+            const requestId = req.params.id;
+
+            // check if the request is exist or not before deleteting the request
+            const request = await FRIEND_REQ.findById(requestId);
+
+            if (!request) {
+                return res
+                    .status(404)
+                    .json({ message: "Frined request not found" });
+            }
+
+            await FRIEND_REQ.findByIdAndDelete(requestId);
+            res.status(200).json({ message: "Friend request rejected" });
+        } catch (error) {
+            res.status(500).json({
+                error: error.message,
+                message: "Internal server error",
+            });
+        }
+    },
+
+    getAllFriends: async (req, res) => {
+        try {
+            const userId = req.userId;
+            const user = await USERS.findById(userId).populate("friends");
+
+            const friends = user.friends.map((frnd) => ({
+                _id: frnd._id,
+                displayName: frnd.displayName,
+                userName: frnd.userName,
+                createdAt: frnd.createdAt,
+                friends: frnd.friends,
+            }));
+
+            res.status(200).json({ friends });
+        } catch (error) {
+            res.status(500).json({
+                error: error.message,
+                message: "Internal Server Error",
             });
         }
     },
